Add DocumentKind::LatestUpload DI (#3693)
## Describe your changes

We're going to start storing the last uploaded document image in order to be able to decrypt it on the dashboard when it's been uploaded

## Issue: <FP-???>

## Caveats

I don't love that we can no longer use these macros for FromStr, Display, and EnumIter - the DocumentKind code is now pretty messy. The reason it doesn't bother me too much: I'd ideally like to move OCR data and image data into very similar patterns to this `LastUpload` where we have only a handful of DocumentKinds:
- `Verified(IdDocKind, DocumentSide)` or just `Image(...)`
- `LatestUpload(IdDocKind, DocumentSide)`
- `OcrData(IdDocKind, OcrDataKind)`

Another caveat: may one day change this approach to show all document uploads instead of just the latest. Hiding the latest_upload DI from docs so we can break it in the future