[FP-4308] Alpaca review action (#3663)
## Describe your changes

A possible way to handle reviews in the AlpacaKYC workflow. 🤔 Not too
sure how I feel about this- sorta getting at the existential question of
how much we want to encapsulate inside vs outside workflows.

In this approach, in `/entities/{fp_id}/decisions`:
- We move the current code for writing obds/reviews to a common
`review::save_review_decision` util
- We query for if the user is in a workflow (right now just queries for
latest wf by creation time but we should add a concept of only 1 active
wf a time).
- If there is one, then we run the `ReviewCompleted` action on that
workflow. AlpacaKYC is currently the only workflow expecting this
action. For now, it simply calls the same `review::save_review_decision`
function and transitions the workflow to `complete`. There may later be
other alpaca specific stuff we need to do at the same time, but for now
I don't think there is.
- If there is no active WF or the WF isn't expecting the
`ReviewCompleted` action then we just gracefully defer to the existing
code (now in `review::save_review_decision`). I think long term what we
probably want is a concept of a single active WF at any given point in
time and if there is an active workflow and its not expecting
`ReviewCompleted` then we error.

Alternatively, we could just keep the route in tact and then after it
makes updates to OBD/manualreview/onboarding, we then materialize the WF
and tell it the review was completed. But this would break the PG writes
and the transition change into different transactions.
Still many open questions and directions we could take this Workflow
idea so open to any and all ideas here! 😄


## Issue: FP-4308

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->