Prepare for making a new IncodeVerificationSession for each IdDoc (#3717)
## Describe your changes

When we have a workflow to re-upload an id doc, we'll need to be able to make a new incode verification session to track the state of the state machine for the new document upload session.

This rms scoped_vault_id since it was just denormalized - we never actually read it, and identity_document_id has all the scoped vault information we need.