[FP-4426] Serialize onboarding requirement for onboarding process (#3789)
## Describe your changes

When the workflow needs acknowledgement from the client that user input is finished, serialize an ob requirement for it.

## Caveats

For now, this won't affect normal KYC flows since they don't create a workflow in the DB yet. even when they do, it will be fine since the POST /authorize endpoint also meets the Process requirement, so it doesn't matter when we land the frontend change to handle this

## Test Plan

Added to the integration test for redo_kyc that we deal with the process requirement