Fix email bug in collect kyc data machine (#3794)
## Describe your changes

this was a tricky bug that had just never occurred before - if we enter the collect KYC data machine where the requirement is met (and we just want to show the confirm screen) but we didn't enter through the identify flow, we won't have an email on file and will instead have a scrubbed email.

this updates the email page to use the same logic to determine if we want to show it as other data

## Caveats

I could see an argument for wanting to continue showing the email page when we have a bootstrapped email - this would be different logic from the other pages, but might be worthwhile since the confirm page never shows the email.

Let me know if you think this should be more thought through for bootstrap :)