Use ScrubbedPiiString in response structs (#3779)
## Describe your changes

For structs for vendor responses, we should be using `ScrubbedPii*`
instead of `Pii*` fields. By default, the `Pii*` fields are Serialize
and will serialize PII. Since we are writing a plaintext version of this
on Vres, we want the PII fields to be scrubbed from that. For our latest
vendor additions (incode, middesk), we forgot this and were using `Pii*`
fields in the response structs. This PR quick fixes these.

I think a better solution here though is to have only 1 Pii wrapped type
per type of data. This type should not be default serializable in any
way (no Display, Serialize, Diesel ToSql, etc) (having a Debug impl
which just prints <SCRUBBED> is chill tho). Any callsite that needs to
serialize the plaintext held within these, should have to do an explicit
`leak` call. So instead of just implicitly/silently leaking when we
serialize structs for API calls, we should have those call sites need to
explicit show they are leaking. Not entirely sure what this would look
like for `Serialize` but presumably we could write some sort of
`#[SerializeWithLeak]` type of macro that provides a
`.leak_and_serialize()` method on the struct.

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->