Leak incode watchlist hits for cip serialization (#3790)
## Describe your changes

With fixing the watchlist response struct to now have `ScrubbedPii`
instead of `Pii` , we now can't simply serialize the response struct and
forward in the Alpaca CIP because various fields will be scrubbed. I
couldn't find a clever way to deal with this short of implementing a
custom Serializer which I ejected from pretty quickly. So an easier but
uglier solution here is to make `Leaked*` versions of the response
structs that have `Pii` instead of `ScrubbedPii`. Then in the alpaca CIP
endpoint, we take our `Vec<Hit>` and `.leak` this into a
`Vec<LeakedHit>` which we can now throw in `serde_json::to_value` and
get a non scrubbed json blob for Alpaca.

Would love to know if there is a better short term solution here. OTOH,
I think its fairly likely that Alpaca will come back to us saying they
want the watchlist records in a certain format and then we might have to
write a custom serializer anyway so not sure this is worth spending too
much time to solve well atm

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->