[FP-4383] Support redo doc workflows (#3811)
## Describe your changes

Very v0 of workflow for doc collection. Surprisingly worked out of the box :P

## Caveats

In the future, we need to stop showing the collect KYC data requirement as met in this document workflow

Need to make some changes to show the document in dashboard too

## Test Plan

Ran through it locally! It surprisingly works out of the box.