[FP-4057] Don't create global fingerprints for fixture vaults (#3783)
## Describe your changes

This fixes a really weird behavior where
1. a vault is created with a fixture number
2. we identify a vault using an email from that fixture number - we used to create global fingerprints for the emails on fixture vaults
3. in identify/verify after the challenge, we look up the vault via phone number and don't find anything, so we make a new vault

now, we don't make global fingerprints for any data on a fixture vault. this limits them to only being identified by ob configs inside the tenant

## Caveats
I'm normally not a big fan of adding these denormalized one-off columns like this, but it drastically simplified the codepath for checking if a vault is a fixture. we could use it in a few other places too

## Test plan

Integration tests