[FP-4440] Expand can see permissions (#3829)
## Describe your changes

Before this, for portable users, we would only ever show that attributes existed in a vault _if_ they were explicitly requested by the ob config. This won't work for future progressive onboarding use cases where, for ex, we might ask to collect a document after onboarding.

NOTE: this isn't controlling decryption permissions - this is just controlling which checkboxes are even visible in the dashboard. The idea now is if the piece of data is visible per DataLifetime rules, we show it in the dashboard