Add method to view scopes to TenantAuth (#3830)
## Describe your changes

In a future PR, we'll need to see the auth scopes and essentially check if a guard is met by the authed principal's scopes without erroring.

## Caveats

Client tenant auth is really weird here. The codepaths in question don't use this, so it's not too bad. But maybe a sign that client tenant auth should live elsewhere