Update experian reason codes to contain a top level identity attribute matching code (#3838)
## Describe your changes

Simplify ssn match codes to not produce all the individual data
attribute codes, and just produce a top level code for each signal scope
to be used in rules


* ssn codes speak in terms of Address and Name
* address result codes speak in terms of individual attributes (street
number, zip code)

currently we assume an "Address match" SSN code maps to the individual
attributes matching (zip matches, street matches etc) but this is a
little odd since we aren't directly getting this information from
experian. I think it's probably okay to provide these codes, but 1) it's
unclear tenants need this 2) makes consolidating the set of reason codes
across vendors difficult.

This PR just makes it so, no matter which type of code we are handling,
we always produce a single "top level" matching code for the signal
scope which we can use in rules

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

updated unit tests