[FP-4459] Support idempotency-id in POST /users (#3836)
## Describe your changes

This was a request from grid - when they make a user in footprint, they want to protect against the case where they crash before saving the new fp_id in their database. So, they want `POST /users` to be idempotent per internal user ID that they have.

This will be a very common pattern for anybody migrating an existing user base into footprint, so i added first class support for it with a header provided to `POST /users`

## Caveats

Will add docs if we like

## Test Plan

integration tests