[FP-4439] Only serialize relevant requirements for the workflow (#3866)
## Describe your changes

Now that we'll have different kinds of workflows, we only want to show the requirements relevant to the currently running workflow.

For ex, if i am in a collect document workflow, i don't want to even serialize that there is a met `collect_kyc_data` requirement - we don't want the user to confirm their KYC data during a collect document workflow

This is built in a way so if we add a new type of workflow, a match statement will fail to compile until we list out the requirements necessary. If we add a new kind of requirement, a match will also fail to compile in `get_requirement_inner`, hinting that you should also update the logic that computes the kinds of requirements needed for a workflow

## Test plan

Will write full recollect doc workflow integration tests soon