[FP-4460 Use backend-powered decryptable_attributes (#3874)
## Describe your changes

Before this PR, the frontend used to duplicate what has become increasingly complex logic to determine which fields are decryptable. It's a function of (1) the entity in question and its populated fields, whether it's portable, which attributes were authorized by the ob, etc and (2) the authed principal at the tenant and the role that they have.

Now, all this frontend logic can just be a simple check if the list of `decryptableAttributes` includes a given di :)

## Test Plan

Before this PR, id documents collected via a collect document workflow wouldn't be decryptable in the dashboard. Now they are!