Alpaca CIP fixes (#3882)
## Describe your changes

A few small fixes to the Alpaca CIP endpoint from Brittany's response:
- For the `watchlist` and `identity` sections, properly set the overall
`result` to `"consider"` if any inner result is `"consider"`
- The dummy Vres we make for the watchlist hit case has both OFAC and
Adverse media hits but we were only producing the OFAC reason code. Adds
the adverse media wat
chlist code too

Still todo: For the document section we also similarily need overall
`result`'s to be a function of inner result's. We also need a plan for
how we will set the data
 comparison fields 



## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->