[FP-4415] Verify that uploaded document satisfies ob config requirements (#3929)
## Describe your changes

This finally enforces the restrictions put in place on documents via the ob config. If the ob config requires us-only docs or a specific doc type, we now error when the document is being uploaded.

## Caveats

Worth noting, we verify the client-provided country code and doc type match the doc request, and then we also later verify that the incode-returned doc type and country code match the client-provided values. I'd like to clean this up by generally getting rid of DocvData, but it's used in a handful of different pockets around the code

## Test Plan

Integration test of an ob config with restrictions