Poll task by kind, add webhook task (#3914)
## Describe your changes

For now, we still don't have a proper daemon/task worker setup. So in the interim, we will manually from application logic prompt the execution of webhook Tasks. We only want these callsites to kick off the execution of webhook tasks (and not other arbitrary tasks) and only do a small number. This PR extends `Task::poll` query to enable for polling of Task's only of a certain kind. This will allow us to query for just `FireWebhookTask`'s.

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

Unit test