Document CIP values from FRCs in alpaca cip endpoint
## Describe your changes

populate document related CIP fields with FRCs. 

there are some TODOs and not precise 1-1 mappings

one minor Q that comes up:

for now, we're good since all docs go to review. but if we ever start auto-approving, the risk is that there's an incode test that doesn't run or an FRC that doesn't get produced for whatever reason and we get a verified document, but one of the data breakdown fields is consider. it's going to be tricky to get this precisely in sync but that's a problem for future us I guess

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->