[FP-4648] Try to catch 500 on idempotency ID constraint violation (#4030)
## Describe your changes

My hypothesis is the 500s are coming from rapidfire requests with the same idempotency ID - we do an explicit DB read before inserting to check for this error case up front. So, we'd expect to catch this with our DB read most of the time. i'm not sure why they're happening so frequently, though...