[FP-4653] Don't show trigger for non-portable vaults (#4040)
## Describe your changes

Small tweak i realized: we error on the backend when a workflow is triggered for a non-portable vault. in order to support this, we need to support step-up in the identify machine from a token that has 0 scopes (since the phone numbers on non-portable vaults aren't verified).

Also, i only show these for person vaults (not businesses)

## Test plan
locally:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/95e1fde8-da4e-41c3-b756-0eea67525a23/image.png)

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/5a2273c9-46df-44f8-b8ee-081b0d9f67ea/image.png)