[FP-4638] Represent vault-only users as None status in the dashboard (#4042)
## Describe your changes

As we start to bridge the gap between non-portable vaults (created via API) and portable vaults (created via bifrost), we want the mental model of them to branch less in dashboard too. Now, instead of being a "vault" user, these users are simply a user that has no status - the onboarding hasn't been started, we're not waiting for a decision, and it isn't passed or failed.

In the future, we'll have the ability to run KYC on these API vaults, which will update this status.

## Caveats

Right now, it is true that (status == None) IFF (user is created via API) (except in one rare case). Eventually, as we start to allow KYCing vaults made via API, we will lose the information on whether a user was created via API...

we're going to start creating timeline events for users made via API, which will definitely help. but maybe we'll also want a little asterisk or flag on vaults that were "imported" or "created via API"

## Test Plan
locally

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/85b55572-81d6-429e-ae21-7412f28262b2/image.png)

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/b035b7c8-a750-411e-98e4-3ac87c389834/image.png)