[FP-1206] Show decrypted API key value as soon as it is created (#4071)
## Describe your changes

This has long been on my TODO list :)

The normal GET endpoint has every API key encrypted by default, but when you make an API key, the response actually shows the revealed API key value. We were previously throwing this response away - now i just prepend it to the list of API keys

## Caveats

This is precarious because it both infers the sort order of results from the backend (if we put new API keys at the bottom of the list, for ex, this would have weird behavior) and the cache query key.

In reality, we want a map of api key ID -> decrypted value that doesn't go away until the page refreshes. But, this does the trick for now

## Test Plan

tested locally

https://github.com/onefootprint/monorepo/assets/21066341/844b16d9-2752-4f6c-915a-6ae6ecf8bebb