Vault integrity check api (#4115)
## Describe your changes

At a high-level, this PR creates a new endpoint `/entities/{fp_id}/vault/integrity` which computes a signed hash of vault data in the enclave. This is a requested feature from Grid who will use this to run a batch job to verify that the footprint vault stores the same contents of the data that was pushed.

At a slightly lower-level: this PR modifies our decrypt functionality to instead be a "functional" decrypt using data transforms. This is neat because the actual hmac happens inside the enclave! More so, this sets up for supporting other `fn_decrypt` operations like vault proxy filter functions.

## Caveats
An integrity operation currently creates "decrypt" access logs, we should instead update our access logs to support a functional decrypt (not only a decrypt)

- [ ] Update access logs for functional decrypts.

Also: should we make a separate IAM permission for functional decrypts?

- [ ] Determine proper IAM for integrity endpoint.

## Test Plan
Adds an integration test.