Flexcar vaulting + KYC scripts (#4200)
## Describe your changes

Scripts and test data for vaulting + KYCing the ~31k sample of users
from Flexcar!

`upsert_vaults.py`:
- Reads the Flexcar provided CSV and create a Footprint user for each
row and vaults the data
- Users are created idempotently using our lovely `x-idempotency-id` and
we set this equal to `Order_ID` which is a unique identifier provided by
Flexcar in their csv data
- PII fields from the csv data are mapped into our `id.*` identifiers
where possible and the rest of the data is stored as
`custom.<col_name_in_csv>`
- Outputs a raw log of all requests and responses and also a final log
of (Order_ID, fp_id) pairs

`run_kyc.py`:
- Reads in a file of (Order_ID, fp_id) pairs and runs /kyc for each
fp_id.
- `/kyc` is by nature idempotent- we only create 1 onboarding/KYC
workflow per fp_id and don't currently support doing multiple KYCs for 1
NPV
- Outputs a raw log fo all requests and responses and also a final log
of (Order_ID, fp_id, /kyc response) rows

`fake_data.csv`
- A dummy file of data in the same shape as the Flexcar provided csv but
that contains no PII/real data
- There are ~50 possible combinations of field null-ness in the provided
Flexcar CSV and this mimics all of those, making this a decently robust
test set for hitting edge cases we'll hit with the real data dump.
- Data is faked/random but for categorical data, the same categories are
used as in the real Flexcar csv

The main quirks with the data to overcome were:
- Address is missing for ~2k rows and that is a hard Idology/Experian
requirement so these rows cannot be KYC'd
- Other rows are missing a combination of other PII fields. This means
we can't use the `must_collect_data` CDO's that OBC's are currently by
default created with or we hit validation errors with missing data
requirements. To overcome this, we need to manually set
`must_collect_data = {name,partial_address}` (or perhaps even `{}`) and
`can_access_data = {email,name,partial_address,phone_number,dob,ssn9}`
in PG for the OBC we use for this exercise.
- The csv has both `MiddleAndSurName` and `LastName` where the latter is
always the suffix of the former. Right now, I vault `id.last_name` as
`MiddleAndSurName` which will vault both the middle and last name and
then `LastName` is just stored under `custom.LastName`
- No rows have SSN, so performance with vendors will probably be dicey
:/
- One field in the CSV is `Chargeback_$` and `custom.Chargeback_$`
throws a validation error because we don't allow symbols in custom key
names. I just rename this to `custom.Chargeback_usd`
- `Phone_Number` is not in e164 format in the csv, so we just prepend
`+` to the provided phone numbers before vaulting

## Issue: <FP-???>

## Caveats

right now params for the script are defined as constants at the top of
the files, but could probably change this to proper CLI args if thats
nicer

## Test Plan

<!--- How did you test your changes? :) -->