risk_signal vendors -> vendor_api (#3999)
## Describe your changes

Replaces `vendors` column in `risk_signal` with `vendor_api`. We
currently are only effectively writing RS's with `vendors` being a list
with a single element. This migrates this into an explicit singular
`vendor_api` and forces call sites creating RS's to adhere to this new
concept of RS's being 1:1 with vendorapi/vres's.
`vendor_api` probably doesn't have to exist longterm on `risk_signal`
but its a nice denormalization and will help with the backfill for
`verification_result_id`

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->