RiskSignal backwards compatible query (#4099)
## Describe your changes

Now onboarding_decision_id is Optional on RiskSignal but we are still
always writing RS's with this set. To start writing RS's without this
set (and instead with verification_result_id set), the immediate hurdle
is to not break the existing functionality of `RiskSignal::
list_by_onboarding_decision_id` which has a few dependencies. This PR
makes this query backwards compatible so that it will still return the
latest RS's associated with a particular OBD either:
- if a RS is a "legacy" RS written with onboarding_decision_id, then we
simple check that as we have been
- if a RS is a "new" RS written with verification_result_id, then we
join that via `onboarding_decision_verification_result_junction` and use
that to filter on the supplied onboarding_decision_id

Longterm, we do not what this query to be reliant on this junction table
and we will instead move over to querying RS's by taking the latest per
risk_signal_group.kind. But this branch preserves functionality until
those later branches are merged in as well

This PR itself doesn't change any callsites to write RS's with vres_id
(ie no non-test usages of NewVres) but prepares for the next PR to do
this

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->