Run  `Check valid db migrations`  before integration tests (#4239)
## Describe your changes

A common problem ive run into a few times is when a PR introduces db
writes that are incompatible with the latest migrations `down.sql`. The
classic example is in PR #1 you make a field optional, and then in PR #2
you start writing null to that field. But CI will run integration tests
first, and then run `Check valid db migrations` after which will hit
issues when it sees the null values the tests wrote (since down.sql is
trying to set the field back to not null)

@agrinman  lemme know if this is kosher or if theres a better solution 

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->