Fix RiskSignal queries (#4452)
## Describe your changes

- Fixes bug causing `GET /entities/{fp_id}/risk_signals/{signal_id}` to
sometimes error. We are now not directly writing onboarding_decision_id
on risk_signal's and instead are associating risk_signal's with a
verification_result_id and a risk_signal_group, which has
scoped_vault_id. Amends the `RiskSignal::get` query to instead query via
risk_signal_group.scoped_vault_id and to filter on hidden=false.
- Fixes bug causing 'duplicate' risk signals to be shown. We are now
writing risk signals for both waterfall KYC vendors but we only set
hidden=false for the chosen vendor. We weren't filtering on this in
`RiskSignal::list_by_onboarding_decision_id` which this PR corrects

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->