[Fix] Update base input sizes to rem (#4532)
We moved the font sizes to rems, but inputs still relied on fixed `px`
values for that. In most cases, this is unnoticeable, but it could lead
to some inconsistencies if the HTML base font size in root is absent in
a project. In this case, some units will remain constant in `px` and the
rems will adapt to a different `px` value than desired.