[FP-5279] Decrypt records for multiple users in batch (#4634)
## Describe your changes

Using the utils defined below, issues requests to the enclave for multiple different e_private_keys in one request. This will massively improve performance for vault proxy operations for multiple users

## Test plan
After this is deployed, i have a performance test case that I've tested on old versions of this code. Will look at the trace for the new version of the code