Write 1 DI per workflow (#4622)
## Describe your changes

Currently, we get_or_create DI's based on sv_id, di.kind. This PR
changes this so we get_or_create base on wf_id, di.kind. This will now
make it so we have 1 DI per workflow.
In followups, we can go a step further and have multiple DI's per
workflow and use di.kind to dileneate different states/groups of vreqs
within a workflow.

KYB is still not migrated to Workflow's so it retains legacy logic of
get_or_create'ing based on sv_id

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->