StytchFingerprintEvent (#4691)
## Describe your changes

Adds a `stytch_fingerprint_event` table that explicitly store the 5 fingerprints Stytch produces and associate those with (session_id, uv_id, sv_id). At the moment, when we receive the device fingerprint payload from Stytch, we just store this in a Vres as the usual unstructured/redacted-or-encrypted json blob. When we actually need to make decision/produce feature/signals based on these fingerprints, presumably the v0 of that would entail querying for historical instances of those fingerprints we have already seen (ie in the absence of a fancy feature pipeline doing this). This PG table would equip us to do that.

## Issue: <FP-???>

## Caveats

I _think_ these fingerprints don't constitute PII and don't need to be encrypted in PG but curious for other folks' thoughts here

## Test Plan

<!--- How did you test your changes? :) -->