Save workos auth method in auth sessions (#4749)
## Describe your changes

We sadly have to save this inside the session since the session can be converted into a real tenant rb auth at any time later.

So, we map the workos auth methods to our own internal representation. In a future PR, we will start enforcing that the auth method matches the allowed methods